// frontend-angular/src/frontend-client/model/restaurantDto.ts
/**
 * Restaurant Import API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RestaurantDto { 
    id: string;
    name: string;
    telephone: string;
    address: string;
    url: string;
    lastCrawled?: string;
    lastEdited?: string;
}